requirements.txt
certifi>=2024.0.0
cffi>=1.16.0
curl_cffi>=0.14.0
pycparser>=1.21
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
psycopg[binary]>=3.1.18
# Tuta 邮箱支持（登录/解密）
cryptography>=42.0.0
argon2-cffi>=23.1.0
# 可选：验证码识别（混合方案）
numpy>=1.26.0
opencv-python>=4.9.0.80
# 可选：无痕打开支付页需要 playwright（pip install playwright && playwright install chromium）
# playwright>=1.40.0
